modif de p en int32

+++ c-skeleton/headers/combinaison.h
@@ -5,7 +5,7 @@
 #include "../headers/point.h"
 
 // Fonction qui calcule la factorielle d'un entier
-uint64_t combinaison(uint64_t p, uint32_t k);
+uint64_t combinaison(int32_t p, uint32_t k);
 
 // Fonction qui génère la combinaison suivante de k éléments pris parmi n
 // Elle renvoie NULL si aucune combinaison suivante n'est possible
@@ -21,6 +21,6 @@ uint64_t *next_comb(uint64_t comb[], uint64_t n, uint64_t k);
  * @param p Nombre total de vecteurs dans l'ensemble d'entrée
  * @return Tableau de pointeurs vers les combinaisons générées, ou NULL en cas d'erreur
  */
-point_t ***generate_combinations(point_t **vectors, uint64_t n, uint64_t k, uint32_t p);
+point_t ***generate_combinations(point_t **vectors, uint64_t n, uint64_t k, int32_t p);
 
 #endif //COMBINATION_H
